+++ .gitignore
@@ -1,2 +1,3 @@
 .DS_STORE
 __pycache__
+.env
