+++ deploy/Dockerfile
@@ -7,14 +7,14 @@ RUN pip install poetry==1.2.2
 RUN poetry config virtualenvs.create false
 
 # Copying requirements of a project
-COPY pyproject.toml poetry.lock /app/src/
-WORKDIR /app/src
+COPY pyproject.toml poetry.lock /app/
+WORKDIR /app
 
 # Installing requirements
 RUN poetry install
 
-# Copying actuall application
-COPY . /app/src/
+# Copying actual application
+COPY . /app/
 RUN poetry install
 
 CMD ["/usr/local/bin/python", "-m", "power_systems_data_api_demonstrator"]
