+++ power_systems_data_api_demonstrator/__main__.py
@@ -6,7 +6,7 @@ from power_systems_data_api_demonstrator.settings import settings
 def main() -> None:
     """Entrypoint of the application."""
     uvicorn.run(
-        "power_systems_data_api_demonstrator.web.application:get_app",
+        "power_systems_data_api_demonstrator.src.application:get_app",
         workers=settings.workers_count,
         host=settings.host,
         port=settings.port,
