+++ power_systems_data_api_demonstrator/src/api/docs/__init__.py
@@ -1,0 +1,4 @@
+"""Routes for swagger and redoc."""
+from power_systems_data_api_demonstrator.src.api.docs.views import router
+
+__all__ = ["router"]
