+++ power_systems_data_api_demonstrator/src/api/grid_node/__init__.py
@@ -1,0 +1,4 @@
+"""Dummy model API."""
+from power_systems_data_api_demonstrator.src.api.grid_node.views import router
+
+__all__ = ["router"]
