+++ power_systems_data_api_demonstrator/src/api/monitoring/__init__.py
@@ -1,0 +1,4 @@
+"""API for checking project status."""
+from power_systems_data_api_demonstrator.src.api.monitoring.views import router
+
+__all__ = ["router"]
