+++ power_systems_data_api_demonstrator/src/api/monitoring/views.py
