+++ power_systems_data_api_demonstrator/src/api/router.py
@@ -1,10 +1,6 @@
 from fastapi.routing import APIRouter
 
-from power_systems_data_api_demonstrator.web.api import (
-    docs,
-    grid_node,
-    monitoring,
-)
+from power_systems_data_api_demonstrator.src.api import docs, grid_node, monitoring
 
 api_router = APIRouter()
 api_router.include_router(grid_node.router, prefix="/gridNode", tags=["gridNode"])
