--- power_systems_data_api_demonstrator/web/api/docs/__init__.py
@@ -1,4 +1,0 @@
-"""Routes for swagger and redoc."""
-from power_systems_data_api_demonstrator.web.api.docs.views import router
-
-__all__ = ["router"]
