--- power_systems_data_api_demonstrator/web/api/monitoring/__init__.py
@@ -1,4 +1,0 @@
-"""API for checking project status."""
-from power_systems_data_api_demonstrator.web.api.monitoring.views import router
-
-__all__ = ["router"]
